PoC: transpile to CommonJS allows to eliminate transformers,
but we are left with some ugly replacements of `require` with
`await require`; maybe it could be possible to perform them
in a `after` transformer, it would require further work.

The advantage of this approach is in lower maintenance burden
(once in TypeScript version is released); the disadvantage is that
the transpiled code looks less like the original making debugging
harder for end users.